styling of view-board and modal
