+++ client/src/App.css
@@ -1,8 +1,149 @@
 /* App.css */
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
 
 body {
   font-family: Arial, sans-serif;
-  color: #000;
-  font-size: 16px;
-  margin: 20px;
+  background-color: #f0f0f0;
+  color: black;
+}
+
+/* Board Details */
+.board-deatil {
+  padding: 20px;
+  max-width: 1200px;
+  margin: auto;
+}
+
+/* Buttons */
+button {
+  background-color: #e0e0e0;
+  color: black;
+  border: none;
+  padding: 8px 12px;
+  margin: 5px;
+  cursor: pointer;
+  border-radius: 4px;
+  font-weight: bold;
+}
+
+button:hover {
+  background-color: #c0c0c0;
+}
+
+/* Card Grid & Cards */
+.cards-grid {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 16px;
+  justify-content: center;
+  margin-top: 20px;
+}
+
+.card {
+  background: white;
+  border: 1px solid #ddd;
+  padding: 12px;
+  border-radius: 6px;
+  width: 250px;
+  text-align: center;
+  box-shadow: 0 2px 5px rgba(0,0,0,0.1);
+}
+
+.card img {
+  width: 100%;
+  height: 200px;
+  object-fit: cover;
+}
+
+/* Modals (Add Board & Add Card) */
+.modal-overlay {
+  position: fixed;
+  inset: 0;
+  background-color: rgba(128,128,128,0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 999;
+}
+
+.modal-content {
+  background: white;
+  padding: 20px;
+  border-radius: 8px;
+  width: 90%;
+  max-width: 400px;
+  color: black;
+}
+
+.modal-content label,
+.category-filter label {
+  display: block;
+  margin-top: 10px;
+  font-weight: bold;
+  color: black;
+}
+
+.modal-content input,
+.modal-content select,
+.category-filter select {
+  width: 100%;
+  padding: 8px;
+  margin-top: 5px;
+  font-size: 14px;
+  color: black;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  background: white;
+}
+
+.category-filter {
+  margin-bottom: 10px;
+}
+
+.modal-buttons {
+  margin-top: 20px;
+  display: flex;
+  justify-content: flex-end;
+  gap: 10px;
+}
+
+.modal-buttons button {
+  cursor: pointer;
+}
+
+/* GIF Results */
+.gif-results {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+  justify-content: center;
+  margin: 10px 0;
+}
+
+.gif-results img {
+  width: 100px;
+  height: 100px;
+  object-fit: cover;
+  cursor: pointer;
+}
+
+/* Responsive */
+@media (max-width: 768px) {
+  .cards-grid {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .card {
+    width: 90%;
+  }
+
+  .gif-results img {
+    width: 80px;
+    height: 80px;
+  }
 }
